Update graphql metadata names
